.js to .ts(tsx) for index / helpers
