+++ src/helpers/setDebug.ts
@@ -1,4 +1,4 @@
-export default function setDebug (envSetting) {
+export default function setDebug (envSetting: string) {
     if (envSetting === 'dev') {
       return false
     } else {
