+++ src/services/consts.ts
@@ -1,0 +1,16 @@
+export const _APP = "qryn-query";
+export const _HISTORY_ITEM = _APP + "-history-item";
+export const _TIMERANGE_ITEM = _APP + "-time-range-item";
+export const _CHART_ITEM = _APP + "-chart-item";
+export const _LABELS_ITEM = _APP + '-labels-item';
+export const _URL_ITEM = _APP + "-url-item";
+export const _QUERY_ITEM = _APP + "-url-item";
+
+// module.exports = {
+//     _HISTORY_ITEM,
+//     _CHART_ITEM,
+//     _LABELS_ITEM,
+//     _TIMERANGE_ITEM,
+//     _URL_ITEM,
+//     _QUERY_ITEM
+// }
