+++ webpack.config.js
@@ -5,7 +5,7 @@ const webpack = require("webpack");
 
 module.exports = {
     mode: "development",
-    entry: "./src/index.js",
+    entry: "./src/index.tsx",
     output: {
         path: path.resolve(__dirname, "build"),
         filename: "bundled.js",
